Splitted polygon series into type files.
