+++ ts/Series/Polygon/PolygonPoint.d.ts
@@ -1,0 +1,35 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ * Imports
+ *
+ * */
+import type PolygonPointOptions from './PolygonPointOptions';
+import PolygonSeries from './PolygonSeries.js';
+import ScatterPoint from '../Scatter/ScatterPoint.js';
+
+/* *
+ *
+ * Class
+ *
+ * */
+declare class PolygonPoint extends ScatterPoint {
+    public options: PolygonPointOptions;
+    public series: PolygonSeries;
+}
+
+/* *
+ *
+ * Export
+ *
+ * */
+export default PolygonPoint;
