+++ ts/Series/Polygon/PolygonPointOptions.d.ts
@@ -1,0 +1,31 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ * Imports
+ *
+ * */
+import ScatterPointOptions from '../Scatter/ScatterPointOptions.js';
+
+/* *
+ *
+ * Declarations
+ *
+ * */
+export interface PolygonPointOptions extends ScatterPointOptions {
+}
+
+/* *
+ *
+ * Export
+ *
+ * */
+export default PolygonPointOptions;
